Decoupled API story and task creation.
In combination with review #77013, this commit decouples the
creation of tasks and stories. The reason for this is threefold:

1- It's more RESTful. Modifying one resource should not modify
another one.
2- Leaving them coupled doesn't provide a meaningful benefit that
cannot be handled by the client.
3- It frees up our ability to design client flows for different
use cases. For example, the web client might enforce a rule that
is "at least one task per story", which would allow us to expand
the 'create story' dialog to allow a user to create multiple tasks.
Similarly, we can add story creation templates that start with a
number of default tasks.

Change-Id: I6478e78435e66b86b5b527805217bb50a1a8d20e