Added DELETE method for projects, stories, and tasks.
While true deletion of records is a bad idea, we still need an ability to
remove individual items from showing up in searches. This commit aims to
lay the ground work for this, though we'll want to put better permissions
around it.

I could use some help on this - for some reason running the tests dies a
horrible death because it gets confused about the migrations.

Change-Id: Ib0eb2f06553644878fb7bf6ad49a820efcdf22e0