Remove Branch and Milestone legacy tables
Remove the Branch and Milestone legacy tables from the POC. Those will
very likely be implemented in a very different way now.

While it is unused yet, I kept the StoryTag table around, since its
implementation is likely to be quite similar.

In order to delete the milestone_id foreign key in the tasks table, we
need the name of the constraint. So I retroactively named all foreign
keys in table 'tasks' under the autogenerated name for MySQL databases.

Change-Id: Id3e8e12b19046aed243b8f073fe18a98c03011e6