Introducing the event timeline
The api calls will now create events with a 'event_type' set to an
action that was performed.
Regular user comments will now also have their own type.

TimeLine Events have an optional reference to a comment (comment_id). If
comment_id is set then a comment_field will be resolved to bring the
content of the comment along with the event in one response.

The event_info field contains additional information about the event
which now includes:
* task_title
* old/new_status
* old/new_assignee_name

This information shouild be helpfull for UI to render the event w/o
making additional requests to the server-side.

Change-Id: I2ad04dbe47c249ee06d9753a164a597889c810b9