Validate Project and Project Groups names
The fields like Projects' names should be validated.

The special WSME type added to check names for a regex and length.
The pattern allows alphanumeric characters with dots, dashes and
salashes for separators. The name should be at least 3 symbols long and should
not start or end with a separator. No sequential separators are allowed.

Existing projects will be renamed to a temporary safe name. The original
name will be appended to the description.

Change-Id: I2304cdfa130283507b1415053117e83d9516e9be