Added subscriber and publisher modules
For each action, except for creation of a new project, two messages will
be sent to rabbitmq. One with the main resource; resource_id; user_id;
method and the other with the event_id; user_id; a faked method POST and
a faked resource TIMELINE_EVENT

Publisher creates an exchange called 'storyboard' and publishes the
messages to it. Subscriber creates the queues with different binding_keys
which bind themself to the storyboard exchange and start consuming the
messages from the exchange.

The consumed messages will be printed on the console for now which will
be modified later.

Do not merge until storyboard has Rabbitmq running.

Change-Id: Ic4697f79aaab82dadf1fb1ae66f414a90ae28dac