Added a subscription_events table & subscription hander
This table will be populated with the subscriber id, the type of event
that occurred and the event-info in the resources he subscribed to.

A person subscribed to a project-group will be notified of
1. Addition/ removal of projects to that project group.
2. Changes to stories, tasks of any of the projects under the
project-group.

A person subscribed to a project will be notified of
1. Changes to stories, tasks associated with that project.
2. When the project is added or removed from a project-group.

A person subscribed to a story will be notified of
1. Any changes to the tasks associated with that story

A person subscribed to a task will be notified of
1. Any changes to the task.

Added authentication checks for current user in v1/subscription_events.py

Added handle_deletions method in subscriptions_handler.py
This method is to remove any active subscriptions(tasks,project_groups)
on the resources that have been just deleted.

Changed the db migration revision number for subscription_events
as it conflicted with fulltext indexes

Added a method to resolve comment content.

Edit: Removed foreign key constraint on subscriber_id in
subscription_events.py from migrations.

Change-Id: Ie226f652241d22d5cee96876370b5797d5bfa54d