Added subscriber and publisher modules
For each action, except for creation of a new project, two messages will
be sent to rabbitmq. One with the main resource; resource_id; user_id;
method and the other with the event_id; user_id; a faked method POST and
a faked resource TIMELINE_EVENT

Publisher creates an exchange called 'storyboard' and publishes the
messages to it. Subscriber creates the queues with different binding_keys
which bind themself to the storyboard exchange and start consuming the
messages from the exchange.

The consumed messages are used by CR 113016.

Co-Authored-By: Nikita Konovalov <nkonovalov@mirantis.com>

Change-Id: Ia573437302dc2d0b1a68d2343e83f9dd397fac04