Maintain launchpad id's in migration and manage autoincrement.
This patch extracts the launchpad ID during migration, and uses it
as the Story ID in StoryBoard. In order to make sure that we don't
end up in ID conflicts with StoryBoard's autoincrement, and MySQL
will not allow auto_increment to be set to anything lower than
the largest used ID, I've also provided a commandline mechanism
to set the auto-increment value directly during import. This allows
the user to "make room" for launchpad stories.

Change-Id: I36c542c7ecad7161cd1f3570b291c3500c4b2dff