Consolidated functional test data bootstrap
Too much boilerplate data loading, instead we create one database
state for all of our functional tests. In the case of the user tests,
I switched the bit and the ID due to changes in the mock data.

NOTE: This patch exposed a nasty left-join bug in StorySummary,
which I temporarily worked around. A more concrete set of
tests around StorySummary needs to be established.

Change-Id: I1b7fa05af0ab69859a54214831c07fa995fef7b2