Working Directory Management
This patch creates a configurable working directory which storyboard may
use to persist files on disk. Example uses may be file uploads, deferred
processing of emails, storing pid files, and others. The library also
provides an API by which a plugin may request their own, isolated
working directory.

The default configuration is ~/.storyboard, as we have a reasonable
expectation that the current user has the ability to create their own
directories. More advanced installs may want to put this in
/var/lib/storyboard or /opt/storyboard.

Change-Id: I41adefbbba7466ab82240b3c079f1e43e261dfbb