Permit name conflicts in users table.
This patch removes the uniqueness constraint from the username
column in the users table. This change was prompted by Launchpad's
support of duplicate user names, as it seems to index off of openid
instead. Furthermore, in order to simplify resolving a given user's
openid, we now hash based on the input parameters of the openid
request rather than the username.

Fact is that we don't really use the username for anything.
We use the display name, and the openid, however the username seems
to be an attempt at retaining some semblance of a mapping to IRC
handles, which I contend is not the right way handle it, as
there may be a 1-to-n relationship of IRC names.

Change-Id: I0ed99eec8e1f01783ec6efbaad9a99120562fce1