Clean recovery for import.
This patch adds additional duplication checks on all stories,
tasks, and comments, built on the assumption that it's ok for
the import to be destructive. If the importing ID is found,
the record is overwritten, and it is checked for the proper events,
as well as the number of discussion comments found.

Change-Id: I5e305c5d02598c9ac572b84682befcb55db0b4d7