Launchpad Import now imports multi-project bugs.
This patch does two thigns: First it removes the on-disk imported
stories cache, so that a story found for a different project may be
imported and assigned to that project as well. Secondly, it adds
additional filtering to check for story tasks in the associated
project, so that new tasks may also be added.

Also adds more human-readable output.

Change-Id: I49c0a96e25fab41926030e5bd34fdd7fe1b5b7ae