Add API call to return task statuses
/v1/task_statuses will return the list of task statuses to the
webclient. They are still hardcoded as enum in server, but add
the ability to return this enum to the webclient so we get rid
of hardcoding that constants on the web side.
In later steps we could replace the enum in the server for a
proper table that can be maintained by admins, but it has some
implications about how to group tasks in the api, how to make
a difference between special states such as todo, merged, etc...

Change-Id: I16cc7886f69fdefe05808af5035859fd818f3492