Fixed bug in working directories.
The patch that was merged for working path directories included three
bugs: Firstly, the use of ListOpt for the working_directory
configuration. Secondly, it did not properly expand '~' user
references, and instead treated them as actual file paths. Thankfully
this feature is not yet in use, so there should not be any files that
need to be cleaned up on production. Lastly, the working dir was
never set on subsequent runs, as the assignment was blocked in an if
statement.

Change-Id: I112d054d8cb9e76ad38ef3c2b36ca700804238df