Converted application to use utcnow()
This patch fixes inconsistent usage of now() and utcnow() in
storyboard. The default oslo timestamp mixin prescribes the use
of utcnow() as the default method of determining datetimes,
and blending this with the use of now() in many locations resulted
in inconsistent date calculations. While on infra this was not a
problem as both server and database use UTC as their server time,
it becomes a problem on developer systems that use local timezones,
as the default python datetime.now() does not return a datetime
object with a properly set tzinfo.

Change-Id: I7470164f448e09916f84daaf91b68c07a876c501