Corrected type casting in OAuth Response
wsgiref performs header type checking when a response is prepared,
including an explicit call of assert type(val) is StringType.
Using six.text_type here cast our value to unicode, and sadly
StringType in wsgiref is hard-bound to the 'str' method, thus
causing the response construction to fail with a type exception.

This is preventing users from logging in.

Change-Id: I2c32f144a76a5e629204958fc520c4686a327705