Email Templating Engine
This commit adds an email templating engine which permits batch
generation of emails. By default, it requires a text template,
however other text template types may be added to the engine and will
be appended in order. It is contained inside the email
plugin directory, as it is one component of the emailing system.

Change-Id: I90ee44425807e96d2fb3ddc0adf122a54636a266