Working Directory fails consistenly.
get_working_directory has been updated to raise an IOError if it
is not able to create and/or access a working directory. Additional
checks have been added to catch these cases:

- If a plugin tries to execute, and the working directory fails,
it will not execute.
- If the cron plugin is registered, it will confirm that the
working directory is accesible before reporting that it is enabled.
- Working directory no longer attempts to fall back to a
non-configured value, and instead fails cleanly.

Change-Id: I04011cdb1cec24f4621c1ff1f0090a59d56509c8