Singularized resource names.
We inconsistently use resource name plurals and singulars throughout
storyboard, mostly due to some initial confusion during code creation.
The tables are plural, the API endpoints are plural, the resource
names are singular, the subscription records are singular. This patch
moves internal usage of resource names to be 'singular' and match
what is in the subscription table, for eventual conversion to an enum.

Change-Id: I7785a121c0cf9fc7a02b7bdfea2c5ced12750c1a