Created test class for temporary working directories
Unexpected use of get_working_dir() can easily pollute the test
environment, as it makes use of a global WORKING_DIRECTORY parameter
that will remain constant between test invocations and will not
trigger a clean. This patch adds a test case base class that
overrides the configuration defaults for working_directory, and
ensures that they are cleaned up after the tests execute.

Change-Id: I9ad27cc317f6f2512c29a80dac669e5766843a80