Pass the old entity values in the NotificationHook
This change allows in the NotificationHook after method to
access the old state of the entity changed by the API call, via
the state.old_entity_values variable.
We may want to add that to the publish method signature in
a later change, so we get in the events queue a message showing
the old vs new state of the resource changed.

Change-Id: I0f21e512cdd4e2904e6b1f0c2c2aa4070e5ab786