Email SMTP Client
This patch adds a class that acts as a context-managed SMTP client
factory, based on the configuration options provided via the previous
patch. It is expected that any consumer of this method will batch
send all of its messages, as the context handler will automatically
call quit() once the context is exited.

A generic monkey-patched mock smtplib is also provided.

Change-Id: Ifbc5231ea2ecb3fb6b6a9015c9273df72bca4091