Switch from mysql-python to PyMySQL
mysql-python undesireable over PyMySQL for a few reasons. Mainly, it
does not support Py3* and does not work well with eventlet.

There is also a bug in using oslo.db with PyMySQL where reference errors
are not properly detected. This fixed in oslo.db
ab20754db71e55b79b9e71e36ad86d9befc89a92 but we should still work around
this unless global-requirements are bumped for oslo.db.

Also adding pymysql to requirements since it is a pure python library
and having a codebase that can 'just work' is awesome.

Change-Id: I0058193c7cbd329731ade37108614fa6eb19d0f7