OAuth Authorize Scope Validation
This patch checks to ensure that oauth authorization requests
include a 'scope' parameter, and that this parameter is set to
'user'. At this time, scope is not yet fully implemented in
storyboard as we do not have comprehensive ACL's, however as soon
as we add our permissions layer, scope adjudication can be
passed to our ACL layer instead.

Change-Id: I9e4448946ade59b1eaa3c21b662665e52b39bab7