OAuth now validates client ID against a configured list.
This patch adds the valid_oauth_clients configuration option, a list
of string ID's that limits what the client may use to identify itself.
This is a breaking change for downstream consumers, as the domain on
which storyboard is hosted will need to be added to this
configuration option before logins will again be permitted.

Change-Id: I411fdf67b56f5be0d151152471e5a00a75daa51e