Datetime instances from SQLAlchemy now all contain timezone.
One of the complex problems in storyboard has been that none of our
datetime instances contain a timezone, making comparison operations
guesswork at best. This is partially exacerbated by oslo.db's
TimestampMixin, which ensures created_at and updated_at parameters
are set, but does not inform the instance that its underlying
assumption is to store datetime in UTC.

This patch addresses this, by creating our own UTCDateTime SQLAlchemy
type decorator and applying it everywhere a datetime is needed.

Change-Id: Iea89661e4d370f0e8af93d2c033b4c44723a39a4