Added python-dateutil
Inside WSME's parse_isodatetime there's a switch that checks for the
existence of python_dateutil. If this does not exist, it falls back
on its own ISO parsing, which does not support timezones, and causes
havoc in our API. This adds the python_datetime dependency
(BSD License) that triggers this switch.

An unrelated change has been submitted to the WSME project to handle
timezone parsing: https://review.openstack.org/#/c/161013/

Change-Id: Ib1bbe1118cc04f7622d1d0ef3993594d5bd102f7