Removed Process Pool execution.
There appear to be differences in Python2.7 and Python 3.4 that
prevent proper pickling of methods passed to the process pool executor,
meaning that our plugins cannot actually work in python 2.7. As this
component is reasonably well isolated, fixing it so that execute_plugin
works as expected can be deferred in favor of getting the feature
working now.

Change-Id: I4b46a1a348065607d4d9c8f4b8c18a7e7b26c82d