Remove page_size_maximum
This patch removes maximum page size from all API calls. I've
realized that it is not for the API to force the clients to
implement paging. Some clients may find it easier to grab single,
large requests, while others want to filter and page their
requests. There are cases in which result limits are beneficial,
and some in which they are not - it is up to the client to decide
which use case is best for them.

Furthermore, this patch sets the minimum size of a page to 0. This
is to support queries for information only (how much do I have).

Change-Id: Ie6c45fad43756dae762b8f973b30439c3340f558