Remove page_size_default
This patch removes page_size_default. This is done as a followup
to removing page_size_maximum, because the justification for
removing that defines the use case of getting all records. With
page_size_default in place, getting all records is not strictly
possible, as an API consumer must always declare a 'large' page
size in order to ensure they get everything, and even in that
case it's not guaranteed. Thus, making 'limit' an optional
parameter switches the default API behavior to getting all records.

Change-Id: I4a2ffdfbb91dab79bdbfa440e40c1fc8fe62637c