Project Loader creates default branches.
A bug was found in the project load script where newly created projects
were not persisted to the database (and thus granted a unique id)
before that ID was queried to generate the branch. Tests have been
added to isolate and test this case, and the issue itself has been
fixed.

Change-Id: I019d52ee1809e373ab3651095671f52bfd934c06