Fixed scheduler management conflict.
This corrects an intermittent issue in the scheduler, where a race
condition caused by the initialization (which itself runs
initialization by loading it as a cron job) would attempt to update a
plugin job out of order. This has been fixed by explicitly setting
replace_existing in the plugin management.

Change-Id: Id1748d6d202e9310f80da810639a4ff104d7d2d4