Work around "duplicate preferences" issue
This patch works around the issue with duplicate preference
records causing it to seem like updating preferences is broken
on storyboard.openstack.org.

This should be made unnecessary once we figure out why duplicate
preferences are being created in the first place, but will fix
the current production instance in the meantime.

Change-Id: Ieef4fdb7d27598daf3d70ee047b39ee310744a36