Update all affected items in a worklist when an item moves
When an item moves in a worklist, the position of some other items
also needs to change. Previously the webclient handled this by
updating every item in the list with a new position, with an
individual PUT request for each one. This patch renders that behaviour
unnecessary by updating each affected item's position in addition to
the position of the item the PUT request was for.

This also means that the user's view of a lane doesn't affect the
change made to the database when an item is moved. This fixes an
issue where multiple users editing a board without refreshing to
see each other's changes could cause cards to disappear and be
duplicated, and also made concurrent editing of worklists impossible
without refreshing before each change.

Change-Id: Ideef3680f6948cbdd371ebb28a36ef9205754b11