Archive cards instead of deleting them
This patch adds an "archived" field to WorklistItems, which
signifies whether or not the card is archived. It also makes
the delete endpoint for cards set this field to True rather
than deleting the card entirely.

This patch also makes the function for adding new cards to
boards/worklists check if there is a card for the same object
archived in the board/worklist, and if there is restores that
card rather than creating a new one.

Change-Id: Iacb2733a2c647f179c29a7d6449497f65b868860