Fix updating automatic worklists with no filters
If a worklist had items, and was then made automatic, those items are
preserved in the database. If the newly automatic worklist has no
filters defined, and an update is attempted, it would send the empty
list to the API as the new set of items it contains. This would cause
the API to attempt to update the items table in an invalid way, since
worklist.as_dict(omit_unset=True) omits things which are equal to
wtypes.Unset, not things which evaluate to False.

This bug would exist for manual worklists too, if not for the fact that
there are no items to update incorrectly in an empty manual list.

This commit fixes this issue, by checking if the value has anything set
(including a value that would evaluate to False, such as the empty
list) in the same way the omit_unset parameter used to get the dict
later on does.

Change-Id: I9307734adcf0dc01dc446d94e9e9ffad02772458