Fix error in search due to multiple `title` columns in query
Searching for tasks is broken because we attempt to filter a query
with multiple columns called `title` ambiguously (on `title` instead
of `tasks.title`).

This is due to a bug in the version of sqlalchemy_fulltext currently
in use on storyboard.openstack.org (version 0.2). This bug was fixed by
version 0.2.2, but we may as well require 0.2.3 (the current latest
release).

Story: 2000579
Change-Id: Ia8c8c6a35a52188ac8f596fb7fed0eb7adbbea33