Update story.updated_at when tasks are updated
The updated_at field of stories is currently only updated when
the db record of the story itself is updated.  However, it is
more useful if it is also updated when other intrinsic attributes
of the story (regardless of whether they are stored in that table)
are updated as well.  This change adds the behavior for tasks.

Change-Id: I41335a7f7219a6e8ff046812eed8fa24cb86ed66