Add filters to the /v1/tasks/search endpoint
This commit adds almost the same set of parameters that the /v1/tasks
endpoint supports to the /v1/tasks/search endpoint. This allows one
to obtain some search results and then filter them more easily than
before, as well as avoiding a bug when trying to naively search for
tasks without removing the default criteria in the web ui.

Eventually we should probably merge the two endpoints.

Change-Id: Ie65d5b8a23c96369050013face90f15d3a39d84b