Fix install guide MySQL admin instructions
This commit amends the lines about the MySQL password prompt so that
they do not suggest the user input the password in the body of the
command(s) to modify the StoryBoard database.
MySQL prompts the user for a password *after* the user enters
a command, so this makes our instructions more accurate and easier
to follow.

Change-Id: Ifc38ee869c8232e61f542b7a4753f0411c71e237