Handle Branches from Launchpad in Migration
When migrating bugs from launchpad to StoryBoard, there is only
logic in place for master branches. If a bug exists on a branch
other than master, its added to master anyway. This patch gets the
branch info from the bug, checks that the branch exists in the project
and if it doesn't advises on creating the branch and exits. Otherwise
the bug is created.

This patch also cleans up other pep8 errors related to print
statements.

Change-Id: I0c1c8d3ebef80fd0937eaa343fdb26f7aadec67c