Minimise database queries when populating automatic worklists
When populating automatic worklists we currently throw away
queries representing the list of story and task objects to
appear in the worklist, and then later on reacquire this data
using one query per story or task.

This commit makes use of the ability to resolve stories and
tasks from a pre-loaded cache to reuse the queries constructed
when first populating the list to get the story data with no
additional database interaction.

This commit also makes the population of the task or story's
due dates optional, as this doesn't make sense for automatic
worklists (wherein due dates cannot be applied).

Change-Id: I76f29e9b9100693691187cffa64b3e6a02120ccb