Pin os-testr to version 0.8.2
This commit pins the os-testr version to 0.8.2, as more recent changes
to the backend mean that some storyboard modules are not found. In the
future, we should adjust storyboard to work with the more recent
os-testr version, either by changing paths to those modules or
by explicitly pointing to them.

Change-Id: Iad68613d294179548d4ca73d75df699b05e7d0d7