Bump minimum os-testr requirement to 1.0.0
This commit bumps the os-testr version to use 1.0.0. 1.0.0 changes the
internals to use stestr instead of subprocessing testr. A bunch of
incorrect imports in the tests were updated as a well, relying on
relative import paths is finicky at the best of times so to make the
tests work this just points to the path based on the root of the package
instead.

Change-Id: I2aed3f6a9adeac2b7dbe58ee12c0a1485278d56f