Check story permissions when populating automatic worklists
Currently, when populating the items of an automatic worklist,
the permissions set on private stories is not taken into account.

This commit fixes this issue by filtering out stories and tasks
that the user shouldn't be able to see when finding the list
of stories and tasks which match the worklist's filters.

Change-Id: If37be62890db913b428af4e6a94ee21754c6ac56