Pin version of webob to 1.7.4
Pecan depends on webob. We don't currently specify a version,
so Pecan pulls in the latest version of webob. (1.8.0rc1)
There is some incompatibility between our codebase and
the latest version of webob, causing our tests to fail.
This patch pins webob to the last known compatible version,
so that development is not blocked on investigating the
incompatibility.

In the future, we should investigate further so that we
can use up to date python dependencies

Change-Id: I8d4f14c992c18b53888657580a292a21b565de85