Fallback to Natural Language search if syntax is bad
Fulltext search in StoryBoard uses Boolean mode, so some characters
have semantic meaning (for example -). It is possible to search for
something such as `js-`, which will currently raise a large number
of errors.

This commit fixes this by falling back to Natural Language mode if
there is an error when using Boolean mode. This lets searches for
things like `js-` be done successfully, without mangling the user's
search term (other syntax errors are possible, so its best to not
just assume they are searching for a single "word").

Change-Id: I99f388d6ab8223206856446f103499678bc7adea
Story: 2000595
Task: 3010