suppress warnings from wsmeext.sphinxext
When the wsmeext.sphinxext extension is initialized during the
documentation build it registers some autodoc classes that use the
same objtype as built-in types. This causes the associated directives
to be overridden, which emits a warning. That warning prevents us from
turning on warning-is-error in the build. Since there is nothing we
can do about it, we want to ignore it.

Change-Id: Ie74174176d6140467c60ea864dc2df4dc2287604
Signed-off-by: Doug Hellmann <doug@doughellmann.com>