Make notification driver configurable
This commit makes the notification driver configurable. Currently there
is only one notification backend available, rabbit/pika, but to support
emitting notifications to firehose.openstack.org in the future a MQTT
backend will be added. This commit lays to groundwork for doing this.
Note this only effects the notification publisher, the subscriber calls
will be updated in the future. (there is only one use of the subscriber
currently so that is updated manually to directly use the new path)

The external api for registering a pecan notification hook remains the
same, but the underlying publish call being made is set by the
configuration option. This enables a different publish() method to be
called depending on configuration.

To facilitate this change the current pika notification options are
deprecated in the 'notifications' group and are instead added to the
'pika-notifications' configuration group. This will be removed in the
future. A new 'driver' option is added to the ntofications group to
specify which driver to use which defaults to the current pika driver.

Change-Id: I5235ae4368f8bff13c767b87ced83173d52b5155