do not set OS_TEST_TIMEOUT in tox.ini
Do not set the variable in tox.ini, because that overrides any value
the user is trying to pass on the command line or from their
environment. A default is already set in .testr.conf using shell
syntax that supports using the value from the user if there is one.

Change-Id: I5ff1880655e927ea596217154958fb95857f5649
Signed-off-by: Doug Hellmann <doug@doughellmann.com>