only try to reset the db user password if it needs to be
If the user sets MYSQL_ROOT_PW then presumably the database is already
configured with that password and we don't need to run the admin
command to change it. The command was failing on my system but when I
removed the line the rest of the commands in the script worked fine.

Change-Id: Id7c41efb2f2457254997ae39626f79f7356e8608
Signed-off-by: Doug Hellmann <doug@doughellmann.com>