do not allow story id to change via payload
There is already an explicit check to prevent changing a task through
a URL referring to the wrong
story. storyboard.tests.api.test_tasks.TestTasksNestedController.test_update_error
is failing with sqlite when the story id of a task is modified, but it
fails with the wrong error (404 instead of 400). Update the controller
to test the case explicitly and return 400.

Change-Id: Iabbc9cc733cc96fe77b721d28ee350a9beab7613
Signed-off-by: Doug Hellmann <doug@doughellmann.com>