Use charset=utf8mb4 in pymysql DBURIs
To make full use of the new 4-byte character set in columns for
MySQL, connections to the database need to also specify
charset=utf8mb4. The easiest way to do that is by tacking it on as a
variable in the DBURI string. Do this both in the test migration
script and in the documented sample and test configs.

Change-Id: Icf5f74543f6d062cafe2166d384f37f46394e964