Support quoted project names in GET requests by name
Currently the webclient quotes the project name before sending the
request, which causes a 404 if there is a / in the project name since
the API doesn't unquote the name before searching for it.

This commit makes the API less opinionated by unquoting provided project
names before searching for them, which should make it less frustrating
to use.

Change-Id: I9a0158231f54809b90198875712ac0cbd543a254