skip some alembic migrations for sqlite
SQLite does not support many of the DDL instructions to modify
tables. We only use it for development and unit tests, so we don't
need to support it fully in the alembic migrations.

This patch updates the tests to use sqlalchemy's create_all() function
to initialize the sqlite database, and then skips the migrations or
portions of migrations that would not work for sqlite.

This has 2 benefits: creating the database in the unit tests is faster
and it should simplify any work we need to do to remove constraints
from tables or otherwise change the database schema, since those
changes will only need to be run for mysql tests and not the sqlite
tests.

Change-Id: I6658596ca3687595596d09428f97f9a82f9062ac
Signed-off-by: Doug Hellmann <doug@doughellmann.com>