accept tag values in the body as well as URL
The JS client passes arguments to the PUT call in the URL so we need
the tags argument to the put() method. The python client passes
arguments encoded in the body of the request, so we need the body
argument. Merge the two sets of values so that inside the function we
only have to process one list.

Change-Id: I21a82321fabbe4f0d0a5db5a9db5dda8f65f21bb
Signed-off-by: Doug Hellmann <doug@doughellmann.com>