Fix up a few requirements
The upper cap on oslo.db and sqlalchemy cause very old versions of both
to be installed. The oslo.db one is particularly bad as it includes a
'def async' which does not work with python3 (fascinated how the tox
py35 tests are working)

eventlet and sqlalchemy-migrate were added a while ago to solve
build-time dependency issues in transitive deps. They are no longer
needed.

Change-Id: I964d87408573008c7b7cca6ce0110a910429b9b8