Pin oauthlib<3 for now
The recent oauthlib 3.0.0 release seems to have regressed
authentication, raising "InvalidRedirectURIError: (invalid_request)
Invalid redirect URI." in
oauthlib.oauth2.rfc6749.grant_types.base._handle_redirects() on
returning from the OpenID provider. This is an emergency pin to <3
until we're able to suss out whether this is a bug in storyboard's
authentication client implementation or a legitimate regression in
oauthlib.

Change-Id: Iad4e035d88a0cd335c6ef68263bf7e044a192513