Fixes the Python 3 dependencies and iterator
A. Fixes depedencies of Python 3:

Following the documentation about Python 3 requirements.
It was needed to change in the file ./requirements.txt
replacing "python-openid" with "python3-openid".

B. Fixing Python 3 compatibility in the migration script:

In order to fix the migration script to run on Python 3
I changed the line #24 on ./migrate/launchpad/reader.py,
changing the iterator.

Also, I added  in the file "./requirements.txt" a comment
to use the library "launchpadlib" in order to run
the migrate script.

Change-Id: Iedcd283ade1ea62fdc5777dfe7b1ae5916db3480