Use ColumnElements instead of strings in migration
Using strings to reference columns in `select()` is no longer
supported, and real ColumnElement objects should be used instead.
Previously this was a warning, but now it is an error which breaks
our database migrations with latest SQLAlchemy.

This commit fixes the only place we do this in our code.

Change-Id: I3134089525f76330c455fff83ab0ec90e4e20f0d