Pin tox-py35 jobs to ubuntu-xenial
We use tox-py35 from zuul-jobs directly since we don't use the global
openstack/requirements constraints. However, this means that our
tox-py35 tests are now running on Bionic, which doesn't have Python
3.5.

This commit sets the nodeset for the tox-py35 job to ubuntu-xenial,
to continue not using the constraints but also allow the tests to
actually run again.

Change-Id: I7c4208521b25303d257cb22a84c3841d0b5e8c09